+++ client/src/App.js
@@ -13,7 +13,7 @@ import UserEdit from './components/UserEdit/UserEdit';
 
 import EstablishmentProfile from "./components/EstablishmentProfile/EstablishmentProfile";
 import AddUserToEstablishment from './components/AddUserToEstablishment/AddUserToEstablishment.jsx';
-import CardEstablishment from "./components/CardEstablishment/CardEstablishment";
+import EstablishmentDetail from "./components/EstablishmentDetail/EstablishmentDetail";
 
 
 function App() {
@@ -31,7 +31,7 @@ function App() {
         <Route exact path="/useredit" component={UserEdit} />
         <Route exact path="/establishmentprofile" component={EstablishmentProfile} />
         <Route exact path="/addUserToEstablishment" component={AddUserToEstablishment}/>
-        <Route exact path="/establishment/1" component={CardEstablishment}/>
+        <Route exact path="/establishment/1" component={EstablishmentDetail}/>
         
 
       </div>
